+++ src/main/java/me/jinjjahalgae/presentation/api/ContractController.java
@@ -3,10 +3,10 @@ package me.jinjjahalgae.presentation.api;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import me.jinjjahalgae.domain.contract.usecase.create.dto.CreateContractRequest;
-import me.jinjjahalgae.domain.contract.usecase.update.ContractUpdateRequest;
+import me.jinjjahalgae.domain.contract.usecase.update.dto.ContractUpdateRequest;
 import me.jinjjahalgae.domain.contract.usecase.create.dto.CreateContractResponse;
-import me.jinjjahalgae.domain.contract.usecase.get.detail.ContractDetailResponse;
-import me.jinjjahalgae.domain.contract.usecase.get.list.ContractListResponse;
+import me.jinjjahalgae.domain.contract.usecase.get.detail.dto.ContractDetailResponse;
+import me.jinjjahalgae.domain.contract.usecase.get.list.dto.ContractListResponse;
 import me.jinjjahalgae.domain.contract.enums.ContractStatus;
 import me.jinjjahalgae.domain.contract.usecase.create.CreateContractUseCase;
 import me.jinjjahalgae.domain.contract.usecase.get.detail.GetContractDetailUseCase;
